requirements.txt
langchain>=0.1.0
langchain_community>=0.0.20
langchain_openai>=0.0.8
langgraph>=0.0.20
openai>=1.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
qdrant-client>=1.6.0
diffusers>=0.25.0
torch>=2.0.0
transformers>=4.36.0
Pillow>=10.0.0
pydantic>=2.0.0
google-cloud-translate==3.12.0 